Refactored all the network codes. Put them in the APIService and called them on the EpisodesController via fetchEpisodes func.
