+++ Podcast/Podcast/API/APIService.swift
@@ -8,6 +8,7 @@
 
 import Foundation
 import Alamofire
+import FeedKit
 
 class APIService {
     
@@ -15,6 +16,26 @@ class APIService {
     
     //singleton
     static let shared = APIService()
+    
+    func fetchEpisodes(feedUrl: String, completionHandler: @escaping ([Episode]) -> ()) {
+        let secureFeedUrl = feedUrl.contains("https") ? feedUrl : feedUrl.replacingOccurrences(of: "http", with: "https")
+        
+        guard let url = URL(string: secureFeedUrl) else { return }
+        let parser = FeedParser(URL: url)
+        parser?.parseAsync(result: { (result) in
+        print("Successfully parse feed:", result.isSuccess)
+        
+        if let err = result.error {
+        print("Failed to parse XML Feed:" , err)
+        return
+        }
+        
+        guard let feed = result.rssFeed else { return }
+        let episodes = feed.toEpisodes()
+        completionHandler(episodes)
+
+        })
+    }
     
     func fetchPodcasts(searchText: String, completionHandler: @escaping ([Podcast]) -> ()) {
         print("Searching for podcast...")
